+++ Sweetsy/Details Page/DetailsViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  DetailsViewModel.swift
+//  Sweetsy
+//
+//  Created by Chetan Dhowlaghar on 5/30/24.
+//
+
+import Foundation
